refactor ModelOptions into options.rs
